requirements.txt
# Core ML
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.0
xgboost==2.0.3
lightgbm==4.3.0

# AWS SDK
boto3==1.34.34
botocore==1.34.34
sagemaker==2.207.0

# Serialization
pyyaml==6.0.1

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
